Remove the use of list elements for non-lists
This changes `<a2j-template>` to use `<section>` elements instead of `<li>` elements.

Unfortunately I was not able to add tests for this change. I wanted to use FuncUnit’s [drag](https://funcunit.com/docs/FuncUnit.prototype.drag.html) method for testing the drag & drop functionality, but I believe it might expect to be used when dragging an element into another element, whereas this component reorders elements in the same parent. It might be possible to add tests for this in the future with more investigation.

Closes https://github.com/CCALI/a2jdeps/issues/52