Update production to npm release 7.1.0 (#60)
* #125 restore user progress (#42)

* html-entities package added

* for extracting visited pages from answers model

* Updated test html-entities

* minor changes to test case

Co-authored-by: Anubhav Dhingra <anubhav.dhingra@launchventures.co>

* remove unused dependancy

* 7.0.19

* updated scope for decodeHTMLEntities

* 7.0.20

* #193 test assemble error (#44)

* syntax update

* bug fix

Co-authored-by: Anubhav Dhingra <anubhav.dhingra@launchventures.co>

* 7.0.21

* 117 conditional logic dat styling (#45)

* styling change for conditional form

* updated stylesheet

* cleanup

Co-authored-by: Anubhav Dhingra <anubhav.dhingra@launchventures.co>

* WIP: template header footer duplicate names (#47)

Co-authored-by: Anubhav Dhingra <anubhav.dhingra@launchventures.co>

* add stripHtml() util to string.js (#50)

* 7.0.22

* author136 - add name tag to unassigned fields (#48)

* add name tag to unassigned fields

* unique id with unidentified fields

* linting cleanup

* counter for unassigned variables

Co-authored-by: Anubhav Dhingra <anubhav.dhingra@launchventures.co>
Co-authored-by: mitch <mitch@mukungfu.com>

* Communicate “page breaks” via more than just styles #51

Previously, the page break element was not perceivable to screen readers because it rendered no text.

Now, some text will be in the DOM for screen readers to read, while the visual representation of the page break has not changed.

Closes https://github.com/CCALI/a2jdeps/issues/51

* Move testee to its own npm script

This allows the tests to be run without running the linter, which is sometimes useful during development.

* Add custom headers and footers to the generated document

Previously, a template’s custom header and footer were not part of the generated HTML preview at all.

Now, the headers and footers are included in the SSR page when the `renderTemplateFootersAndHeaders` attribute is `true`. This will allow us to conditionally turn on the headers and footers when DAT renders the HTML preview but not when it’s generating the PDF.

This adds a number of attributes to one of the existing templates because they are required for the template to successfully render; see this issue for more info: https://github.com/CCALI/a2jdeps/issues/55

Part of https://github.com/CCALI/a2jdeps/issues/53

* Remove the use of list elements for non-lists

This changes `<a2j-template>` to use `<section>` elements instead of `<li>` elements.

Unfortunately I was not able to add tests for this change. I wanted to use FuncUnit’s [drag](https://funcunit.com/docs/FuncUnit.prototype.drag.html) method for testing the drag & drop functionality, but I believe it might expect to be used when dragging an element into another element, whereas this component reorders elements in the same parent. It might be possible to add tests for this in the future with more investigation.

Closes https://github.com/CCALI/a2jdeps/issues/52

* 7.1.0

* 7.1.1-0

* Update README.md

* add textindent plugin to ckeditor (#63)

* 7.1.1

Co-authored-by: AnubhavDhingra <anubhavdhingra95@gmail.com>
Co-authored-by: Anubhav Dhingra <anubhav.dhingra@launchventures.co>
Co-authored-by: Chasen Le Hara <git@chasenlehara.com>
Co-authored-by: Tobias Nteireho <tobias@cali.org>
Co-authored-by: John Mayer <jmayer@cali.org>